rename Id page
